extract getPagePaths, getURLsFrom functions
towards basic dynamics pt. 1 (#124)
